Change application icon for the Clone Project
The application icon for the Clone Project has been updated. In the .fsproj file, the application icon was previously set to 'CloneProject.ico'. This has been replaced with 'icon-code-compare.ico'. The same change was reflected for the file icon in PNG format as well. The change of icon was made to better represent the intention and functionality of the project.